Refactored, removed CoreData support
