+++ SyEuler/Supporting/Assets.xcassets/AppIcon.appiconset/Contents.json
