+++ SyEuler/Supporting/Assets.xcassets/Contents.json
