+++ SyEuler/Supporting/Info.plist
