+++ SyEuler/Supporting/Logger.swift
