+++ SyEuler/Supporting/SyEulerApp.swift
@@ -1,0 +1,17 @@
+//
+//  SyEulerApp.swift
+//  SyEuler
+//
+//  Created by Wayne Mock on 12/29/20.
+//
+
+import SwiftUI
+
+@main
+struct SyEulerApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}
