+++ SyEuler/Views/ContentView.swift
@@ -1,0 +1,21 @@
+//
+//  ContentView.swift
+//  SyEuler
+//
+//  Created by Wayne Mock on 12/29/20.
+//
+
+import SwiftUI
+
+struct ContentView: View {
+    var body: some View {
+        Text("Hello, world!")
+            .padding()
+    }
+}
+
+struct ContentView_Previews: PreviewProvider {
+    static var previews: some View {
+        ContentView()
+    }
+}
